refactor: move title stuff from gridView to separate file
